Conflicts:
	galaxy-all/galaxy-dal-dao/src/main/java/com/galaxy/dal/activity/mapper/ActivityJoinedUsersMapper.java
	galaxy-all/galaxy-dal-dao/src/main/resources/com/galaxy/dal/activity/mapper/ActivityJoinedUsersMapper.xml
	galaxy-all/galaxy-dal-dao/src/main/resources/spring/data-access.properties
	galaxy-all/galaxy-front-all/galaxy-front-service/src/main/java/com/galaxy/service/activity/ActivityService.java
	galaxy-all/galaxy-front-all/galaxy-front-service/src/main/java/com/galaxy/service/activity/impl/ActivityServiceImpl.java
	galaxy-all/galaxy-front-all/galaxy-front-web/src/main/java/com/galaxy/front/web/rest/card/CardController.java

+++ galaxy-all/galaxy-dal-domain/src/main/java/com/galaxy/dal/domain/activity/Activity.java
@@ -25,7 +25,7 @@ public class Activity extends BaseDomain {
 
 	private String refUrl;
 
-	private Double price;
+	private Float price;
 
 	private String tags;
 
@@ -134,10 +134,10 @@ public class Activity extends BaseDomain {
 	public void setRefUrl(String refUrl) {
 		this.refUrl = refUrl;
 	}
-	public Double getPrice() {
+	public Float getPrice() {
 		return price;
 	}
-	public void setPrice(Double price) {
+	public void setPrice(Float price) {
 		this.price = price;
 	}
 	public String getTags() {
