+++ galaxy-all/galaxy-front-all/galaxy-front-web/src/main/java/com/galaxy/front/web/activity/controller/PostModel/CreateModel.java
@@ -20,7 +20,7 @@ public class CreateModel {
 	private String detail;
 	private String ticket_name;
 	private int ticket_num;
-	private Double ticket_price;
+	private Float ticket_price;
 	private int optionsRadios;
 	private Long catId1;
 	private Long catId2;
@@ -137,11 +137,11 @@ public class CreateModel {
 		this.ticket_num = ticket_num;
 	}
 
-	public Double getTicket_price() {
+	public Float getTicket_price() {
 		return ticket_price;
 	}
 
-	public void setTicket_price(Double ticket_price) {
+	public void setTicket_price(Float ticket_price) {
 		this.ticket_price = ticket_price;
 	}
